Move sign() method to MemberIdBundle.sign()
