+++ src/lib/signature.spec.ts
@@ -1,15 +1,9 @@
 import { jest } from '@jest/globals';
 import { type Message, RrSet, SecurityStatus } from '@relaycorp/dnssec';
 import { AsnParser, AsnSerializer } from '@peculiar/asn1-schema';
-import { Null, type OctetString } from 'asn1js';
-import {
-  Attribute,
-  ContentInfo,
-  SignedData as SignedDataSchema,
-  EncapsulatedContent,
-  type SignerInfo,
-} from '@peculiar/asn1-cms';
-import { setMilliseconds, subSeconds } from 'date-fns';
+import { Null } from 'asn1js';
+import { ContentInfo } from '@peculiar/asn1-cms';
+import { subSeconds } from 'date-fns';
 import { Attribute as PkijsAttribute } from 'pkijs';
 
 import { generateMemberIdFixture } from '../testUtils/veraStubs/memberIdFixture.js';
@@ -32,10 +26,10 @@ import { DnssecChainSchema } from './schemas/DnssecChainSchema.js';
 import { SignatureBundleSchema } from './schemas/SignatureBundleSchema.js';
 import { SignedData } from './utils/cms/SignedData.js';
 import Certificate from './utils/x509/Certificate.js';
-import { CMS_OIDS, VERAID_OIDS } from './oids.js';
+import { VERAID_OIDS } from './oids.js';
 import { SignatureMetadataSchema } from './schemas/SignatureMetadataSchema.js';
 import VeraidError from './VeraidError.js';
-import { sign, verify } from './signature.js';
+import { verify } from './signature.js';
 import CmsError from './utils/cms/CmsError.js';
 import { generateTxtRdata } from './dns/rdataSerialisation.js';
 import { MemberIdBundle } from './memberIdBundle/MemberIdBundle.js';
@@ -58,239 +52,15 @@ const MEMBER_ID_BUNDLE = serialiseMemberIdBundle(
   DNSSEC_CHAIN_SERIALISED,
 );
 
-const SIGNATURE_BUNDLE_SERIALISED = await sign(
+// Create a signature bundle for testing
+const bundleInstance = MemberIdBundle.deserialise(MEMBER_ID_BUNDLE);
+const SIGNATURE_BUNDLE_SERIALISED = await bundleInstance.sign(
   PLAINTEXT,
   SERVICE_OID,
-  MEMBER_ID_BUNDLE,
   MEMBER_KEY_PAIR.privateKey,
   datePeriod.end,
   { startDate: datePeriod.start },
 );
-
-describe('sign', () => {
-  test('Malformed member Id bundle should be refused', async () => {
-    const malformedBundle = arrayBufferFrom('malformed');
-
-    await expect(async () =>
-      sign(PLAINTEXT, SERVICE_OID, malformedBundle, MEMBER_KEY_PAIR.privateKey, datePeriod.end),
-    ).rejects.toThrowWithMessage(VeraidError, 'Member id bundle is malformed');
-  });
-
-  test('Version should be 0', async () => {
-    const signatureSerialised = await sign(
-      PLAINTEXT,
-      SERVICE_OID,
-      MEMBER_ID_BUNDLE,
-      MEMBER_KEY_PAIR.privateKey,
-      datePeriod.end,
-    );
-
-    const { version } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-    expect(version).toBe(0);
-  });
-
-  test('DNSSEC chain should be attached', async () => {
-    const signatureSerialised = await sign(
-      PLAINTEXT,
-      SERVICE_OID,
-      MEMBER_ID_BUNDLE,
-      MEMBER_KEY_PAIR.privateKey,
-      datePeriod.end,
-    );
-
-    const { dnssecChain } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-    expect(Buffer.from(AsnSerializer.serialize(dnssecChain))).toStrictEqual(
-      Buffer.from(DNSSEC_CHAIN_SERIALISED),
-    );
-  });
-
-  test('Organisation certificate should be attached', async () => {
-    const signatureSerialised = await sign(
-      PLAINTEXT,
-      SERVICE_OID,
-      MEMBER_ID_BUNDLE,
-      MEMBER_KEY_PAIR.privateKey,
-      datePeriod.end,
-    );
-
-    const { organisationCertificate } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-    expect(Buffer.from(AsnSerializer.serialize(organisationCertificate))).toStrictEqual(
-      Buffer.from(orgCertificateSerialised),
-    );
-  });
-
-  describe('Signature', () => {
-    function getSignedData(contentInfo: ContentInfo) {
-      expect(contentInfo.contentType).toStrictEqual(CMS_OIDS.SIGNED_DATA);
-      return AsnParser.parse(contentInfo.content, SignedDataSchema);
-    }
-
-    test('Plaintext should be signed with specified private key', async () => {
-      const signatureSerialised = await sign(
-        PLAINTEXT,
-        SERVICE_OID,
-        MEMBER_ID_BUNDLE,
-        MEMBER_KEY_PAIR.privateKey,
-        datePeriod.end,
-      );
-
-      const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-      const signedData = SignedData.deserialize(AsnSerializer.serialize(signature));
-      await signedData.verify(PLAINTEXT);
-      const memberCertificate = Certificate.deserialize(memberCertificateSerialised);
-      expect(signedData.signerCertificate!.isEqual(memberCertificate)).toBeTrue();
-    });
-
-    test('Member certificate should be attached', async () => {
-      const signatureSerialised = await sign(
-        PLAINTEXT,
-        SERVICE_OID,
-        MEMBER_ID_BUNDLE,
-        MEMBER_KEY_PAIR.privateKey,
-        datePeriod.end,
-      );
-
-      const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-      const { certificates } = getSignedData(signature);
-      const attachedCertsSerialised = certificates!.map((cert) =>
-        Buffer.from(AsnSerializer.serialize(cert)),
-      );
-      expect(attachedCertsSerialised).toContainEqual(Buffer.from(memberCertificateSerialised));
-    });
-
-    test('Plaintext should be detached by default', async () => {
-      const signatureSerialised = await sign(
-        PLAINTEXT,
-        SERVICE_OID,
-        MEMBER_ID_BUNDLE,
-        MEMBER_KEY_PAIR.privateKey,
-        datePeriod.end,
-      );
-
-      const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-      const { encapContentInfo } = getSignedData(signature);
-      expect(encapContentInfo.eContent).toBeUndefined();
-    });
-
-    test('Plaintext should be attached if requested', async () => {
-      const signatureSerialised = await sign(
-        PLAINTEXT,
-        SERVICE_OID,
-        MEMBER_ID_BUNDLE,
-        MEMBER_KEY_PAIR.privateKey,
-        datePeriod.end,
-        { shouldEncapsulatePlaintext: true },
-      );
-
-      const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-      const { encapContentInfo } = getSignedData(signature);
-      expect(encapContentInfo.eContent).toBeInstanceOf(EncapsulatedContent);
-      const encapsulatedContentSerialised = encapContentInfo.eContent!.any!;
-      const encapsulatedContentAsn1 = derDeserialize(encapsulatedContentSerialised) as OctetString;
-      expect(new Uint8Array(encapsulatedContentAsn1.getValue())).toStrictEqual(
-        new Uint8Array(PLAINTEXT),
-      );
-    });
-
-    describe('Metadata', () => {
-      function getSignedAttribute(signerInfo: SignerInfo, attributeOid: string) {
-        const matchingAttribute = signerInfo.signedAttrs!.find(
-          (attribute) => attribute.attrType === attributeOid,
-        );
-        expect(matchingAttribute).toBeInstanceOf(Attribute);
-        return matchingAttribute!.attrValues;
-      }
-
-      function getSignatureMetadata(signature: ContentInfo) {
-        const { signerInfos } = getSignedData(signature);
-        const attributeValues = getSignedAttribute(
-          signerInfos[0],
-          VERAID_OIDS.SIGNATURE_METADATA_ATTR,
-        );
-        expect(attributeValues).toHaveLength(1);
-        return AsnParser.parse(attributeValues[0], SignatureMetadataSchema);
-      }
-
-      test('Service OID should be attached', async () => {
-        const signatureSerialised = await sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_ID_BUNDLE,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-        );
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const { serviceOid } = getSignatureMetadata(signature);
-        expect(serviceOid).toStrictEqual(SERVICE_OID);
-      });
-
-      test('Expiry date should be attached', async () => {
-        const signatureSerialised = await sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_ID_BUNDLE,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-        );
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const { validityPeriod } = getSignatureMetadata(signature);
-        expect(validityPeriod.end).toStrictEqual(datePeriod.end);
-      });
-
-      test('Start date should default to the current time', async () => {
-        const beforeSignatureDate = setMilliseconds(new Date(), 0);
-        const signatureSerialised = await sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_ID_BUNDLE,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-        );
-        const afterSignatureDate = new Date();
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const { validityPeriod } = getSignatureMetadata(signature);
-        expect(validityPeriod.start).toBeBetween(beforeSignatureDate, afterSignatureDate);
-      });
-
-      test('Any explicit start date should be honoured', async () => {
-        const startDate = subSeconds(datePeriod.start, 1);
-        const signatureSerialised = await sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_ID_BUNDLE,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-          { startDate },
-        );
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const { validityPeriod } = getSignatureMetadata(signature);
-        expect(validityPeriod.start).toStrictEqual(startDate);
-      });
-
-      test('Start date after expiry date should be refused', async () => {
-        const invalidExpiryDate = subSeconds(datePeriod.start, 1);
-
-        await expect(async () =>
-          sign(
-            PLAINTEXT,
-            SERVICE_OID,
-            MEMBER_ID_BUNDLE,
-            MEMBER_KEY_PAIR.privateKey,
-            invalidExpiryDate,
-            { startDate: datePeriod.start },
-          ),
-        ).rejects.toThrowWithMessage(
-          VeraidError,
-          'Signature start date cannot be after expiry date',
-        );
-      });
-    });
-  });
-});
 
 describe('verify', () => {
   interface SignatureBundleAttributeSet {
@@ -444,10 +214,10 @@ describe('verify', () => {
       memberCertificateSerialised, // Invalid
       DNSSEC_CHAIN_SERIALISED,
     );
-    const signatureBundle = await sign(
+    const invalidBundle = MemberIdBundle.deserialise(invalidMemberIdBundle);
+    const signatureBundle = await invalidBundle.sign(
       PLAINTEXT,
       SERVICE_OID,
-      invalidMemberIdBundle,
       MEMBER_KEY_PAIR.privateKey,
       datePeriod.end,
       { startDate: datePeriod.start },
@@ -587,10 +357,10 @@ describe('verify', () => {
         subSeconds(datePeriod.start, 2),
         subSeconds(datePeriod.start, 1),
       );
-      const signatureBundleSerialised = await sign(
+      const testBundle = MemberIdBundle.deserialise(MEMBER_ID_BUNDLE);
+      const signatureBundleSerialised = await testBundle.sign(
         PLAINTEXT,
         SERVICE_OID,
-        MEMBER_ID_BUNDLE,
         MEMBER_KEY_PAIR.privateKey,
         signaturePeriod.end,
         { startDate: signaturePeriod.start },
@@ -622,15 +392,15 @@ describe('verify', () => {
         subSeconds(verificationPeriod.start, 1),
         { startDate: datePeriod.start },
       );
-      const memberIdBundle = serialiseMemberIdBundle(
+      const testMemberIdBundle = serialiseMemberIdBundle(
         otherMemberCertificate,
         orgCertificateSerialised,
         DNSSEC_CHAIN_SERIALISED,
       );
-      const signatureBundleSerialised = await sign(
+      const testBundle = MemberIdBundle.deserialise(testMemberIdBundle);
+      const signatureBundleSerialised = await testBundle.sign(
         PLAINTEXT,
         SERVICE_OID,
-        memberIdBundle,
         MEMBER_KEY_PAIR.privateKey,
         datePeriod.end,
         { startDate: datePeriod.start },
@@ -655,10 +425,10 @@ describe('verify', () => {
 
   describe('Plaintext', () => {
     test('Verification should fail if plaintext is attached and passed', async () => {
-      const bundle = await sign(
+      const testBundle = MemberIdBundle.deserialise(MEMBER_ID_BUNDLE);
+      const signatureBundleSerialised = await testBundle.sign(
         PLAINTEXT,
         SERVICE_OID,
-        MEMBER_ID_BUNDLE,
         MEMBER_KEY_PAIR.privateKey,
         datePeriod.end,
         { shouldEncapsulatePlaintext: true },
@@ -666,7 +436,13 @@ describe('verify', () => {
 
       const error = await getPromiseRejection(
         async () =>
-          verify(PLAINTEXT, bundle, SERVICE_OID, datePeriod, dnssecChainFixture.trustAnchors),
+          verify(
+            PLAINTEXT,
+            signatureBundleSerialised,
+            SERVICE_OID,
+            datePeriod,
+            dnssecChainFixture.trustAnchors,
+          ),
         VeraidError,
       );
 
@@ -674,10 +450,10 @@ describe('verify', () => {
     });
 
     test('Verification should fail if plaintext is detached and not passed', async () => {
-      const bundle = await sign(
+      const testBundle = MemberIdBundle.deserialise(MEMBER_ID_BUNDLE);
+      const signatureBundleSerialised = await testBundle.sign(
         PLAINTEXT,
         SERVICE_OID,
-        MEMBER_ID_BUNDLE,
         MEMBER_KEY_PAIR.privateKey,
         datePeriod.end,
         { shouldEncapsulatePlaintext: false },
@@ -685,7 +461,13 @@ describe('verify', () => {
 
       const error = await getPromiseRejection(
         async () =>
-          verify(undefined, bundle, SERVICE_OID, datePeriod, dnssecChainFixture.trustAnchors),
+          verify(
+            undefined,
+            signatureBundleSerialised,
+            SERVICE_OID,
+            datePeriod,
+            dnssecChainFixture.trustAnchors,
+          ),
         VeraidError,
       );
 
@@ -695,10 +477,10 @@ describe('verify', () => {
 
   describe('Valid result', () => {
     test('Plaintext should be taken from bundle if attached', async () => {
-      const bundle = await sign(
+      const testBundle = MemberIdBundle.deserialise(MEMBER_ID_BUNDLE);
+      const signatureBundleSerialised = await testBundle.sign(
         PLAINTEXT,
         SERVICE_OID,
-        MEMBER_ID_BUNDLE,
         MEMBER_KEY_PAIR.privateKey,
         datePeriod.end,
         { shouldEncapsulatePlaintext: true },
@@ -706,7 +488,7 @@ describe('verify', () => {
 
       const { plaintext } = await verify(
         undefined,
-        bundle,
+        signatureBundleSerialised,
         SERVICE_OID,
         datePeriod,
         dnssecChainFixture.trustAnchors,
@@ -764,15 +546,15 @@ describe('verify', () => {
         datePeriod.end,
         { startDate: datePeriod.start },
       );
-      const memberIdBundle = serialiseMemberIdBundle(
+      const botMemberIdBundle = serialiseMemberIdBundle(
         botCertificate,
         orgCertificateSerialised,
         DNSSEC_CHAIN_SERIALISED,
       );
-      const signatureBundleSerialised = await sign(
+      const botBundle = MemberIdBundle.deserialise(botMemberIdBundle);
+      const signatureBundleSerialised = await botBundle.sign(
         PLAINTEXT,
         SERVICE_OID,
-        memberIdBundle,
         MEMBER_KEY_PAIR.privateKey,
         datePeriod.end,
         { startDate: datePeriod.start },
