+++ src/lib/signature.ts
@@ -1,40 +1,16 @@
 import type { TrustAnchor } from '@relaycorp/dnssec';
 import { AsnParser, AsnSerializer } from '@peculiar/asn1-schema';
-import { ContentInfo } from '@peculiar/asn1-cms';
 import { Certificate as CertificateSchema } from '@peculiar/asn1-x509';
-import type { BaseBlock, Sequence } from 'asn1js';
-import { Attribute } from 'pkijs';
+import type { BaseBlock } from 'asn1js';
 
-import { MemberIdBundleSchema } from './schemas/MemberIdBundleSchema.js';
 import VeraidError from './VeraidError.js';
 import { SignedData } from './utils/cms/SignedData.js';
-import Certificate from './utils/x509/Certificate.js';
 import { SignatureBundleSchema } from './schemas/SignatureBundleSchema.js';
 import { VERAID_OIDS } from './oids.js';
 import { SignatureMetadataSchema } from './schemas/SignatureMetadataSchema.js';
-import { DatePeriodSchema } from './schemas/DatePeriodSchema.js';
-import { derDeserialize } from './utils/asn1.js';
 import { DatePeriod, type IDatePeriod } from './dates.js';
 import { MemberIdBundle } from './memberIdBundle/MemberIdBundle.js';
 import type { SignatureBundleVerification } from './SignatureBundleVerification.js';
-
-function generateMetadata(serviceOid: string, startDate: Date, expiryDate: Date): Sequence {
-  if (expiryDate < startDate) {
-    throw new VeraidError('Signature start date cannot be after expiry date');
-  }
-
-  const metadataSchema = new SignatureMetadataSchema();
-
-  metadataSchema.serviceOid = serviceOid;
-
-  const datePeriod = new DatePeriodSchema();
-  datePeriod.start = startDate;
-  datePeriod.end = expiryDate;
-  metadataSchema.validityPeriod = datePeriod;
-
-  const serialisation = AsnSerializer.serialize(metadataSchema);
-  return derDeserialize(serialisation) as Sequence;
-}
 
 function getMetadata(signedData: SignedData) {
   const metadataAttributeAsn1 = signedData.getSignedAttribute(VERAID_OIDS.SIGNATURE_METADATA_ATTR);
@@ -84,68 +60,6 @@ function getSignaturePeriodIntersection(
     );
   }
   return signaturePeriodIntersection;
-}
-
-async function generateSignedData(
-  plaintext: ArrayBuffer,
-  memberCertificateSchema: CertificateSchema,
-  signingKey: CryptoKey,
-  serviceOid: string,
-  shouldEncapsulatePlaintext: boolean,
-  expiryDate: Date,
-  startDate?: Date,
-) {
-  const memberCertificate = Certificate.deserialize(
-    AsnSerializer.serialize(memberCertificateSchema),
-  );
-  const metadataSchema = generateMetadata(serviceOid, startDate ?? new Date(), expiryDate);
-  const metadataAttribute = new Attribute({
-    type: VERAID_OIDS.SIGNATURE_METADATA_ATTR,
-    values: [metadataSchema],
-  });
-  const signedData = await SignedData.sign(plaintext, signingKey, memberCertificate, [], {
-    extraSignedAttrs: [metadataAttribute],
-    encapsulatePlaintext: shouldEncapsulatePlaintext,
-  });
-  return AsnParser.parse(signedData.serialize(), ContentInfo);
-}
-
-export interface SignatureOptions {
-  readonly startDate: Date;
-  readonly shouldEncapsulatePlaintext: boolean;
-}
-
-export async function sign(
-  plaintext: ArrayBuffer,
-  serviceOid: string,
-  memberIdBundleSerialised: ArrayBuffer,
-  signingKey: CryptoKey,
-  expiryDate: Date,
-  { startDate, shouldEncapsulatePlaintext }: Partial<SignatureOptions> = {},
-): Promise<ArrayBuffer> {
-  let memberIdBundleSchema: MemberIdBundleSchema;
-  try {
-    memberIdBundleSchema = AsnParser.parse(memberIdBundleSerialised, MemberIdBundleSchema);
-  } catch {
-    throw new VeraidError('Member id bundle is malformed');
-  }
-
-  const signedDataSchema = await generateSignedData(
-    plaintext,
-    memberIdBundleSchema.memberCertificate,
-    signingKey,
-    serviceOid,
-    shouldEncapsulatePlaintext ?? false,
-    expiryDate,
-    startDate,
-  );
-
-  const signatureSchema = new SignatureBundleSchema();
-  signatureSchema.version = 0;
-  signatureSchema.dnssecChain = memberIdBundleSchema.dnssecChain;
-  signatureSchema.organisationCertificate = memberIdBundleSchema.organisationCertificate;
-  signatureSchema.signature = signedDataSchema;
-  return AsnSerializer.serialize(signatureSchema);
 }
 
 export async function verify(
